+++ packages/terra-form-select/tests/wdio/select-spec.js
@@ -873,6 +873,20 @@ Terra.describeViewports('Select', ['tiny'], () => {
     });
   });
 
+  describe('Combobox uncontrolled free text clear by escape key press', () => {
+    before(() => {
+      browser.url('/raw/tests/cerner-terra-core-docs/form-select/uncontrolled-combobox');
+    });
+    it('should enter a free text entry in combobox and close list item by pressing escape and clear the free text by pressing again escape', () => {
+      $('[data-terra-select] input').setValue('Black');
+      Terra.validates.element('combobox displayed Add button with typed free text', { selector: '#root' });
+      browser.keys('Escape');
+      Terra.validates.element('combobox dropdown list close by first escape', { selector: '#root' });
+      browser.keys('Escape');
+      Terra.validates.element('combobox cleared entered text by second escape', { selector: '#root' });
+    });
+  });
+
   describe('Combobox Variant - controlled', () => {
     describe('combobox controlled should select an option by click', () => {
       before(() => {
@@ -1540,6 +1554,15 @@ Terra.describeViewports('Select', ['tiny'], () => {
 
       it('should display toggle closed dropdown', () => {
         Terra.validates.element('search close on outside click after toggle open', { selector: '#root' });
+      });
+
+      it('should enter a free text entry in search select and close list item by pressing escape and clear the free text by pressing again escape', () => {
+        $('[data-terra-select] input').setValue('Black');
+        Terra.validates.element('search select displayed displayed typed text', { selector: '#root' });
+        browser.keys('Escape');
+        Terra.validates.element('search select dropdown list close by first escape', { selector: '#root' });
+        browser.keys('Escape');
+        Terra.validates.element('search select cleared entered text by second escape', { selector: '#root' });
       });
     });